strategies/Inference_by_awq/requirements.txt
# AWQ Inference Strategy - Python Dependencies

# Core Dependencies
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# AWQ Quantization
autoawq>=0.1.6

# API Server (optional)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Monitoring & Logging
prometheus-client>=0.18.0
python-multipart>=0.0.6

# Rate Limiting (optional)
slowapi>=0.1.9

# Utils
numpy>=1.24.0
tqdm>=4.66.0
requests>=2.31.0
huggingface-hub>=0.19.0

# Development Tools (optional)
pytest>=7.4.0
black>=23.11.0
flake8>=6.1.0
